+++ include/input/inputFileParser/QMInputParser.hpp
@@ -52,7 +52,7 @@ namespace input
         void parseDispersion(const pq::strings &, const size_t);
         void parseRemoveNetForce(const pq::strings &, const size_t);
 
-        void parseMaceModelSize(const pq::strings &, const size_t);
+        void parseMaceModel(const pq::strings &, const size_t);
         void parseMaceModelPath(const pq::strings &, const size_t);
         void parseMaceQMMethod(const std::string_view &);